Models/requirements.txt
transformers
fastapi
uvicorn
six
pyyaml
python-dotenv
torch
accelerate
